Using `bytes calldata`

+++ contracts/peripheral/Market.sol
@@ -19,8 +19,8 @@ contract Market is IMarket, ERC20, Constants {
         uint64 accumulator;  // interest rate oracle accumulator—32 bits for a UQ16x16, 32 bits for overflow
     }
 
-    int128 constant public k = 126144000 << 64; // Seconds in 4 years, in 64.64
-    int128 constant public g = (999 << 64) / 1000;
+    int128 constant public k = int128(126144000 << 64); // Seconds in 4 years, in 64.64
+    int128 constant public g = int128(uint256((999 << 64)) / 1000); // All constants are `ufixed`, to divide them they must be converted to uint256
     uint256 constant public initialSupply = 1000;
     uint128 immutable public maturity;
